refactor: 관련있는 필드끼리 클래스 관계 수정
